+++ src/commands/hosting-sites-get.ts
@@ -1,4 +1,4 @@
-import Table = require("cli-table");
+const Table = require("cli-table");
 
 import { Command } from "../command";
 import { Site, getSite } from "../hosting/api";