+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Remove call to Cloud Run API and set CPU & concurrency in GCF API instead. (#5605)
 - Fix function deploy retry after quota exceeded bug and increase backoff. (#5601)
 - Fix bug where EVENTARC_CLOUD_EVENT_SOURCE environment variable was correctly set for some functions. (#5597)
+- Adds 2nd gen firestore triggers to firebase deploy (#5592).
+- Adds `database:import` command for non-atomic imports (#5396).